bloque5/chatbot_multimodal/requirements.txt
pandas
google-cloud-aiplatform
streamlit
google-cloud-bigquery
langchain-google-vertexai
vertexai
db-dtypes